more logic shuffling into the config module
