+++ src/fides/api/main.py
@@ -56,7 +56,7 @@ from fides.api.ops.tasks.scheduled.scheduler import scheduler
 from fides.api.ops.util.cache import get_cache
 from fides.api.ops.util.logger import Pii, get_fides_log_record_factory
 from fides.ctl.core.config import FidesConfig, get_config
-from fides.ctl.core.config.utils import check_required_webserver_config_values
+from fides.ctl.core.config.helpers import check_required_webserver_config_values
 from fides.lib.oauth.api.routes.user_endpoints import router as user_router
 
 CONFIG: FidesConfig = get_config()
