+++ src/fides/cli/commands/util.py
@@ -13,6 +13,7 @@ from fides.cli.utils import (
     send_init_analytics,
     with_analytics,
 )
+from fides.ctl.core.config.helpers import create_config_file
 from fides.ctl.core.deploy import (
     check_docker_version,
     check_fides_uploads_dir,
@@ -23,7 +24,6 @@ from fides.ctl.core.deploy import (
     teardown_application,
 )
 from fides.ctl.core.utils import echo_green
-from fides.ctl.core.config.helpers import create_config_file
 
 
 @click.command()
@@ -145,7 +145,7 @@ def up(ctx: click.Context, no_pull: bool = False, no_init: bool = False) -> None
         # Deployment is ready! Perform the same steps as `fides init` to setup CLI
         if not no_init:
             echo_green("Deployment successful! Initializing fides...")
-            config_path = create_config_file(ctx=ctx, fides_directory_location=".")
+            config_path = create_config_file(config=ctx.obj["CONFIG"])
             check_and_update_analytics_config(ctx=ctx, config_path=config_path)
         else:
             echo_green(
