+++ src/fides/ctl/core/config/helpers.py
@@ -1,16 +1,17 @@
 """This module contains logic related to loading/manipulatin/writing the config."""
-from pydantic import BaseSettings
 import logging
-import regex
 import os
-from os import environ
+from os import environ, getenv
 from pathlib import Path
 from typing import Any, Dict, List, Union
-import toml
 
+import toml
 from click import echo
+from pydantic import BaseSettings
+from re import compile as regex
 from toml import dump, load
 
+from fides.ctl.core.utils import echo_red
 
 DEFAULT_CONFIG_PATH = ".fides/fides.toml"
 logger = logging.getLogger(__name__)
@@ -164,6 +165,7 @@ def create_config_file(
 
     return config_path
 
+
 def handle_deprecated_fields(settings: Dict[str, Any]) -> Dict[str, Any]:
     """Custom logic for handling deprecated values."""
 
@@ -203,3 +205,49 @@ def handle_deprecated_env_variables(settings: Dict[str, Any]) -> Dict[str, Any]:
             settings["database"][setting] = val
 
     return settings
+
+
+def check_required_webserver_config_values() -> None:
+    """Check for required env vars and print a user-friendly error message."""
+    required_config_dict = {
+        "app_encryption_key": {
+            "env_var": "FIDES__SECURITY__APP_ENCRYPTION_KEY",
+            "config_subsection": "security",
+        },
+        "oauth_root_client_id": {
+            "env_var": "FIDES__SECURITY__OAUTH_ROOT_CLIENT_ID",
+            "config_subsection": "security",
+        },
+        "oauth_root_client_secret": {
+            "env_var": "FIDES__SECURITY__OAUTH_ROOT_CLIENT_SECRET",
+            "config_subsection": "security",
+        },
+    }
+
+    missing_required_config_vars = []
+    for key, value in required_config_dict.items():
+        try:
+            config_value = getenv(value["env_var"]) or get_config_from_file(
+                "",
+                value["config_subsection"],
+                key,
+            )
+        except FileNotFoundError:
+            config_value = None
+
+        if not config_value:
+            missing_required_config_vars.append(key)
+
+    if missing_required_config_vars:
+        echo_red(
+            "\nThere are missing required configuration variables. Please add the following config variables to either the "
+            "`fides.toml` file or your environment variables to start Fides: \n"
+        )
+        for missing_value in missing_required_config_vars:
+            print(f"- {missing_value}")
+        print(
+            "\nVisit the Fides deployment documentation for more information: "
+            "https://ethyca.github.io/fides/deployment/"
+        )
+
+        raise SystemExit(1)
