+++ src/fides/lib/oauth/api/routes/user_endpoints.py
@@ -17,13 +17,12 @@ from starlette.status import (
     HTTP_404_NOT_FOUND,
 )
 
-from fides.ctl.core.config import FidesConfig
+from fides.ctl.core.config import FidesConfig, get_config
 from fides.lib.models.client import ADMIN_UI_ROOT, ClientDetail
 from fides.lib.models.fides_user import FidesUser
 from fides.lib.models.fides_user_permissions import FidesUserPermissions
 from fides.lib.oauth.api import urn_registry as urls
 from fides.lib.oauth.api.deps import get_db, verify_oauth_client
-from fides.ctl.core.config import get_config
 from fides.lib.oauth.schemas.oauth import AccessToken
 from fides.lib.oauth.schemas.user import (
     UserCreate,
